feat: altera type id

+++ transportadora/src/main/java/br/com/entrega_mais/transportadora/repository/TransportadoraRepository.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TransportadoraRepository extends JpaRepository<Transportadora, Integer> {
+public interface TransportadoraRepository extends JpaRepository<Transportadora, Long> {
 
     Optional<Transportadora> findByCnpj(String cnpj);
-    Optional<Transportadora> findById(long id);
+    Optional<Transportadora> findById(Long id);
     Optional<Transportadora> findById_usuario(String id_usuario);
 
 }
